move fm context stuff to separate file
